weather_forecast/requirements.txt
flask==2.0.1
requests==2.26.0
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
tensorflow==2.9.1
plotly==5.3.1
python-dotenv==0.19.0
joblib==1.1.0
geopy==2.2.0 